fixes bug with hukamnama and few console error related to search-inputs
