+++ src/js/pages/Home/index.js
@@ -94,8 +94,8 @@ export default class Home extends React.PureComponent {
                         placeholder={placeholder}
                         title={title}
                         pattern={pattern}
-                        min={name === 'ang' && 1}
-                        max={name === 'ang' && MAX_ANGS[source]}
+                        min={name === 'ang' ? 1 : undefined}
+                        max={name === 'ang' ? MAX_ANGS[source] : undefined}
                       />
                       <button
                         type="button"
