+++ src/js/util/hukamnama/index.ts
@@ -1,0 +1,1 @@
+export * from './get-hukamnama';
